refactor: move platform packages to internal/
Move darwin/ and freedesktop/ under internal/ to signal they are
not part of the public API. These packages are only consumed by
the cmd/ package and should not be imported by external consumers.
