+++ internal/darwin/browser_service.go
