+++ internal/freedesktop/browser_icon_loader.go
