+++ internal/freedesktop/browser_service.go
