+++ internal/freedesktop/desktop_entry.go
